Separate websocket service
